+++ packages/at_commons/CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 4.0.10
+- fix: Add a "force" variable to enroll_verb_builder to propagate enroll:revoke:force value
 - fix: Deprecate apkam in PkamAuthMode enum
 - chore: deprecate MessageTypeEnum.text
 ## 4.0.9